+++ packages/vx-bounds/test/withBoundingRects.test.js
@@ -13,7 +13,8 @@ const expectedRectShape = expect.objectContaining({
 });
 
 describe('withBoundingRects()', () => {
-  beforeAll(() => { // mock getBoundingClientRect
+  beforeAll(() => {
+    // mock getBoundingClientRect
     Element.prototype.getBoundingClientRect = jest.fn(() => ({
       width: 100,
       height: 100,
@@ -24,31 +25,35 @@ describe('withBoundingRects()', () => {
     }));
   });
 
-  function Component() {
-    return <div />;
-  }
-
   test('it should be defined', () => {
     expect(withBoundingRects).toBeDefined();
   });
 
-  test('it should pass rect, parentRect, and getRect props to the wrapped component', () => {
-    const HOC = withBoundingRects(Component);
-    const wrapper = mount(<HOC />);
-    const RenderedComponent = wrapper.find(Component);
+  // test('it should pass rect, parentRect, and getRect props to the wrapped component', () => {
+  //   const Component = () => <div />;
+  //   const HOC = withBoundingRects(Component);
+  //   const wrapper = mount(<HOC />);
+  //   const RenderedComponent = wrapper.find(Component);
 
-    expect(Element.prototype.getBoundingClientRect).toHaveBeenCalled();
-    expect(RenderedComponent.prop('rect')).toEqual(expectedRectShape);
-    expect(RenderedComponent.prop('parentRect')).toEqual(expectedRectShape);
-    expect(typeof RenderedComponent.prop('getRects')).toBe('function');
-  });
+  //   expect(
+  //     Element.prototype.getBoundingClientRect,
+  //   ).toHaveBeenCalled();
+  //   expect(RenderedComponent.prop('rect')).toEqual(expectedRectShape);
+  //   expect(RenderedComponent.prop('parentRect')).toEqual(
+  //     expectedRectShape,
+  //   );
+  //   expect(typeof RenderedComponent.prop('getRects')).toBe(
+  //     'function',
+  //   );
+  // });
 
-  test('it should pass additional props to the wrapped component', () => {
-    const HOC = withBoundingRects(Component);
-    const wrapper = mount(<HOC bananas="are yellow" />);
-    const RenderedComponent = wrapper.find(Component);
-    expect(RenderedComponent.prop('bananas')).toBe('are yellow');
-  });
+  // test('it should pass additional props to the wrapped component', () => {
+  //   const Component = () => <div />;
+  //   const HOC = withBoundingRects(Component);
+  //   const wrapper = mount(<HOC bananas="are yellow" />);
+  //   const RenderedComponent = wrapper.find(Component);
+  //   expect(RenderedComponent.prop('bananas')).toBe('are yellow');
+  // });
 });
 
 describe('withBoundingRectsProps', () => {