+++ packages/vx-hierarchy/src/hierarchies/Cluster.js
@@ -1,9 +1,15 @@
-import React from "react";
-import cx from "classnames";
-import { Group } from "@vx/group";
-import { cluster as d3cluster } from "d3-hierarchy";
-import DefaultLink from "../HierarchyDefaultLink";
-import DefaultNode from "../HierarchyDefaultNode";
+import React from 'react';
+import PropTypes from 'prop-types';
+import cx from 'classnames';
+import { Group } from '@vx/group';
+import { cluster as d3cluster } from 'd3-hierarchy';
+import DefaultLink from '../HierarchyDefaultLink';
+import DefaultNode from '../HierarchyDefaultNode';
+
+Cluster.propTypes = {
+  root: PropTypes.object.isRequired,
+  children: PropTypes.func,
+};
 
 export default function Cluster({
   top,
@@ -13,6 +19,7 @@ export default function Cluster({
   size,
   nodeSize,
   separation,
+  children,
   linkComponent = DefaultLink,
   nodeComponent = DefaultNode,
   ...restProps
@@ -21,13 +28,29 @@ export default function Cluster({
   if (size) cluster.size(size);
   if (nodeSize) cluster.nodeSize(nodeSize);
   if (separation) cluster.separation(separation);
+
   const data = cluster(root);
-  const links = data.links();
-  const descendants = root.descendants();
+
+  if (!!children) {
+    return (
+      <Group
+        top={top}
+        left={left}
+        className={cx('vx-cluster', className)}
+      >
+        {children({ data })}
+      </Group>
+    );
+  }
+
   return (
-    <Group top={top} left={left} className={cx("vx-cluster", className)}>
+    <Group
+      top={top}
+      left={left}
+      className={cx('vx-cluster', className)}
+    >
       {linkComponent &&
-        links.map((link, i) => {
+        data.links().map((link, i) => {
           return (
             <Group key={`cluster-link-${i}`}>
               {React.createElement(linkComponent, { link })}
@@ -35,7 +58,7 @@ export default function Cluster({
           );
         })}
       {nodeComponent &&
-        descendants.map((node, i) => {
+        data.descendants().map((node, i) => {
           return (
             <Group key={`cluster-node-${i}`}>
               {React.createElement(nodeComponent, { node })}