+++ packages/vx-hierarchy/src/hierarchies/Tree.js
@@ -1,9 +1,15 @@
-import React from "react";
-import cx from "classnames";
-import { Group } from "@vx/group";
-import { tree as d3tree } from "d3-hierarchy";
-import DefaultLink from "../HierarchyDefaultLink";
-import DefaultNode from "../HierarchyDefaultNode";
+import React from 'react';
+import PropTypes from 'prop-types';
+import cx from 'classnames';
+import { Group } from '@vx/group';
+import { tree as d3tree } from 'd3-hierarchy';
+import DefaultLink from '../HierarchyDefaultLink';
+import DefaultNode from '../HierarchyDefaultNode';
+
+Tree.propTypes = {
+  root: PropTypes.object.isRequired,
+  children: PropTypes.func,
+};
 
 export default function Tree({
   top,
@@ -13,6 +19,7 @@ export default function Tree({
   size,
   nodeSize,
   separation,
+  children,
   linkComponent = DefaultLink,
   nodeComponent = DefaultNode,
   ...restProps
@@ -21,13 +28,25 @@ export default function Tree({
   if (size) tree.size(size);
   if (nodeSize) tree.nodeSize(nodeSize);
   if (separation) tree.separation(separation);
+
   const data = tree(root);
-  const links = data.links();
-  const descendants = root.descendants();
+
+  if (!!children) {
+    return (
+      <Group
+        top={top}
+        left={left}
+        className={cx('vx-tree', className)}
+      >
+        {children({ data })}
+      </Group>
+    );
+  }
+
   return (
-    <Group top={top} left={left} className={cx("vx-tree", className)}>
+    <Group top={top} left={left} className={cx('vx-tree', className)}>
       {linkComponent &&
-        links.map((link, i) => {
+        data.links().map((link, i) => {
           return (
             <Group key={`tree-link-${i}`}>
               {React.createElement(linkComponent, { link })}
@@ -35,7 +54,7 @@ export default function Tree({
           );
         })}
       {nodeComponent &&
-        descendants.map((node, i) => {
+        data.descendants().map((node, i) => {
           return (
             <Group key={`tree-node-${i}`}>
               {React.createElement(nodeComponent, { node })}