+++ packages/vx-hierarchy/src/index.js
@@ -1,4 +1,11 @@
-export { default as Tree } from "./hierarchies/Tree";
-export { default as Cluster } from "./hierarchies/Cluster";
-export { default as HierarchyDefaultLink } from "./HierarchyDefaultLink";
-export { default as HierarchyDefaultNode } from "./HierarchyDefaultNode";
+export { default as Tree } from './hierarchies/Tree';
+export { default as Treemap } from './hierarchies/Treemap';
+export { default as Cluster } from './hierarchies/Cluster';
+export { default as Pack } from './hierarchies/Pack';
+export { default as Partition } from './hierarchies/Partition';
+export {
+  default as HierarchyDefaultLink,
+} from './HierarchyDefaultLink';
+export {
+  default as HierarchyDefaultNode,
+} from './HierarchyDefaultNode';