+++ packages/vx-shape/src/shapes/LineRadial.js
@@ -1,7 +1,12 @@
 import React from 'react';
 import cx from 'classnames';
+import PropTypes from 'prop-types';
 import { radialLine } from 'd3-shape';
 import additionalProps from '../util/additionalProps';
+
+LineRadial.propTypes = {
+  innerRef: PropTypes.func,
+};
 
 export default function LineRadial({
   className = '',
@@ -10,6 +15,7 @@ export default function LineRadial({
   defined,
   curve,
   data,
+  innerRef,
   ...restProps
 }) {
   const path = radialLine();
@@ -20,6 +26,7 @@ export default function LineRadial({
   return (
     <g>
       <path
+        ref={innerRef}
         className={cx('vx-line-radial', className)}
         d={path(data)}
         {...additionalProps(restProps, data)}