+++ packages/vx-shape/test/Arc.test.js
@@ -1,21 +1,21 @@
-import React from "react";
-import { shallow } from "enzyme";
-import { Arc } from "../src";
-import { browserUsage } from "../../vx-mock-data";
+import React from 'react';
+import { shallow } from 'enzyme';
+import { Arc } from '../src';
+import { browserUsage } from '../../vx-mock-data';
 
 const ArcWrapper = ({ ...restProps }) =>
   shallow(<Arc data={browserUsage} {...restProps} />);
 
-describe("<Arc />", () => {
-  test("it should be defined", () => {
+describe('<Arc />', () => {
+  test('it should be defined', () => {
     expect(Arc).toBeDefined();
   });
 
-  test("it should have the .vx-arcs-group class", () => {
-    expect(ArcWrapper().prop("className")).toBe("vx-arcs-group");
+  test('it should have the .vx-arcs-group class', () => {
+    expect(ArcWrapper().prop('className')).toBe('vx-arc');
   });
 
-  test("it should contain paths", () => {
-    expect(ArcWrapper().find("path").length).toBeGreaterThan(0);
+  test('it should contain paths', () => {
+    expect(ArcWrapper().find('path').length).toBeGreaterThan(0);
   });
 });