+++ packages/vx-shape/test/AreaClosed.test.js
@@ -1,22 +1,23 @@
-import React from "react";
-import { shallow } from "enzyme";
-import { extent, max } from "d3-array";
-import { AreaClosed } from "../src";
-import { appleStock } from "../../vx-mock-data";
-import { scaleTime, scaleLinear } from "../../vx-scale";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { shallow } from 'enzyme';
+import { extent, max } from 'd3-array';
+import { AreaClosed } from '../src';
+import { appleStock } from '../../vx-mock-data';
+import { scaleTime, scaleLinear } from '../../vx-scale';
 
 const xStock = d => new Date(d.date);
 const yStock = d => d.close;
 
 const fakeXScale = scaleTime({
   range: [0, 100],
-  domain: extent(appleStock, xStock)
+  domain: extent(appleStock, xStock),
 });
 
 const fakeYScale = scaleLinear({
   range: [100, 0],
   domain: [0, max(appleStock, yStock)],
-  nice: true
+  nice: true,
 });
 
 const AreaClosedWrapper = ({ ...restProps }) =>
@@ -28,15 +29,38 @@ const AreaClosedWrapper = ({ ...restProps }) =>
       x={xStock}
       y={yStock}
       {...restProps}
-    />
+    />,
   );
 
-describe("<AreaClosed />", () => {
-  test("it should be defined", () => {
+describe('<AreaClosed />', () => {
+  test('it should be defined', () => {
     expect(AreaClosed).toBeDefined();
   });
 
-  test("it should have the .vx-area-closed class", () => {
-    expect(AreaClosedWrapper().find('path').prop("className")).toBe("vx-area-closed");
+  test('it should have the .vx-area-closed class', () => {
+    expect(
+      AreaClosedWrapper()
+        .find('path')
+        .prop('className'),
+    ).toBe('vx-area-closed');
+  });
+
+  test('it should expose its ref via an innerRef prop', done => {
+    const node = document.createElement('div');
+    const refCallback = n => {
+      expect(n.tagName).toEqual('PATH');
+      done();
+    };
+    ReactDOM.render(
+      <AreaClosed
+        data={appleStock}
+        xScale={fakeXScale}
+        yScale={fakeYScale}
+        x={xStock}
+        y={yStock}
+        innerRef={refCallback}
+      />,
+      node,
+    );
   });
 });